⚡️(frontend) remove auth barrel file to improve code splitting
Auth-related code is used across the codebase and imported by many
different features. The barrel file, while convenient for importing
utilities, was harming code splitting by pulling unrelated modules
into shared chunks.

Verified by running `npm run build:debug` before and after: chunking
and code splitting are better without the barrel.
