--- src/frontend/src/features/auth/index.ts
@@ -1,4 +1,0 @@
-export { useUser } from './api/useUser'
-export { useSyncUserPreferencesWithBackend } from './api/useSyncUserPreferencesWithBackend'
-export { authUrl } from './utils/authUrl'
-export { UserAware } from './components/UserAware'
