+++ src/frontend/src/features/auth/utils/silentLogin.ts
@@ -1,4 +1,4 @@
-import { authUrl } from '@/features/auth'
+import { authUrl } from './authUrl'
 
 const SILENT_LOGIN_RETRY_KEY = 'silent-login-retry'
