+++ src/frontend/src/features/notifications/components/ToastRecordingSaving.tsx
@@ -5,7 +5,7 @@ import { Text } from '@/primitives'
 import { StyledToastContainer, type ToastProps } from './Toast'
 import { HStack } from '@/styled-system/jsx'
 import { useTranslation } from 'react-i18next'
-import { useUser } from '@/features/auth'
+import { useUser } from '@/features/auth/api/useUser'
 import { css } from '@/styled-system/css'
 import { RecordingMode } from '@/features/recording'
