+++ src/frontend/src/features/sdk/routes/CreatePopup.tsx
@@ -1,7 +1,7 @@
 import { useEffect, useMemo } from 'react'
 import { css } from '@/styled-system/css'
 import { generateRoomId, useCreateRoom } from '../../rooms'
-import { useUser } from '@/features/auth'
+import { useUser } from '@/features/auth/api/useUser'
 import { Spinner } from '@/primitives/Spinner'
 import { CallbackIdHandler } from '../utils/CallbackIdHandler'
 import { PopupWindow } from '../utils/PopupWindow'
